Add files via upload
It is created with the help of css,java-script and html. In website is created font-page username and password is same then the website is open and in website Phones and ordered list of various brands phones and link them to the google with the help of anchor tag(<a href=" ">  </a>) .Check then information of that phones in google.